+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ## 1.10.10
-* android: fix platform exception when scanning with includeConnectedDevices (Regression in 1.10.6)
+* android: fix platform exception when scanning with includeConnectedSystemDevices (Regression in 1.10.6)
 * dart: fix characteristic write crashed for negative values (Regression in 1.7.0)
 * dart: fix connectionState should only be concerned with *our apps* connectionState